+++ types/shapes/iter.go
@@ -9,6 +9,15 @@ import (
 // To avoid allocating the slice of indices, the yielded indices is owned by the Iter() method:
 // don't change it inside the loop.
 func (s Shape) Iter() iter.Seq[[]int] {
+	indices := make([]int, s.Rank())
+	return s.IterOn(indices)
+}
+
+// IterOn iterates over all possible indices of the given shape.
+//
+// The iteration updates the indices on the given indices slice.
+// During the iteration the caller shouldn't modify the indices, otherwise it will lead to undefined behavior.
+func (s Shape) IterOn(indices []int) iter.Seq[[]int] {
 	return func(yield func([]int) bool) {
 		if !s.Ok() || s.IsTuple() {
 			return // Iteration completed (vacuously true as no items were yielded)
@@ -36,9 +45,11 @@ func (s Shape) Iter() iter.Seq[[]int] {
 		}
 
 		// Version 1: there are no spatial axes, there is only one value to iterate over.
-		currentIndices := make([]int, rank)
+		for i := range indices {
+			indices[i] = 0
+		}
 		if numSpatialAxes == 0 {
-			yield(currentIndices)
+			yield(indices)
 			return
 		}
 
@@ -48,25 +59,25 @@ func (s Shape) Iter() iter.Seq[[]int] {
 			// This structure simulates an N-dimensional counter for the indices.
 		v2Yielder:
 			for {
-				if !yield(currentIndices) {
+				if !yield(indices) {
 					return // Consumer requested to stop iteration.
 				}
 
-				// Increment currentIndices to the next set of coordinates
+				// Increment indices to the next set of coordinates
 				// (row-major order: the last index changes fastest).
 				for axis := rank - 1; axis >= 0; axis-- {
 					if s.Dimensions[axis] == 1 {
 						// Nothing to iterate at this axis.
 						continue
 					}
-					currentIndices[axis]++
-					if currentIndices[axis] < s.Dimensions[axis] {
+					indices[axis]++
+					if indices[axis] < s.Dimensions[axis] {
 						// Successfully incremented this dimension; no carry-over needed.
 						continue v2Yielder
 					}
 					// The current axis overflowed; reset it to 0 and
 					// continue to increment the next higher-order dimension (carry-over).
-					currentIndices[axis] = 0
+					indices[axis] = 0
 				}
 
 				// If axis is less than 0, all dimensions have been iterated through
@@ -87,21 +98,21 @@ func (s Shape) Iter() iter.Seq[[]int] {
 		slices.Reverse(spatialAxes) // We want to iterate over the last axis first.
 	v3Yielder:
 		for {
-			if !yield(currentIndices) {
+			if !yield(indices) {
 				return // Consumer requested to stop iteration.
 			}
 
-			// Increment currentIndices to the next set of coordinates
+			// Increment indices to the next set of coordinates
 			// (row-major order: the last index changes fastest).
 			for _, axis := range spatialAxes {
-				currentIndices[axis]++
-				if currentIndices[axis] < s.Dimensions[axis] {
+				indices[axis]++
+				if indices[axis] < s.Dimensions[axis] {
 					// Successfully incremented this dimension; no carry-over needed.
 					continue v3Yielder
 				}
 				// The current axis overflowed; reset it to 0 and
 				// continue to increment the next higher-order dimension (carry-over).
-				currentIndices[axis] = 0
+				indices[axis] = 0
 			}
 
 			// That was the last index.