+++ docs/CHANGELOG.md
@@ -13,7 +13,7 @@
 * Package `types/tensors`:
   * Removed dependency to `gopjrt/pjrt` -- otherwise we'll always need to install the C/C++ library.
 * Package `types/shape`:
-  * Added `Shape.Iter()`.
+  * Added `Shape.Iter()` and `Shape.IterOn()`.
 * Package `backend`:
   * `Backend` interface now returns errors instead of panicking.
 * Package `graph`: